Conflicts:
	finalproject/src/main/java/com/example/demo/SecurityConfig.java
	finalproject/src/main/java/com/example/demo/controller/MemberController.java
	finalproject/src/main/java/com/example/demo/dao/MemberDAO.java
	finalproject/src/main/java/com/example/demo/db/DBManager.java
	finalproject/src/main/java/com/example/demo/db/memberMapper.xml
	finalproject/src/main/java/com/example/demo/vo/GoodsDetailVO.java

	0310최현호

+++ finalproject/src/main/java/com/example/demo/vo/MemberVO.java
@@ -15,7 +15,7 @@ public class MemberVO {
 	private String member_addr;
 	private String member_loc;
 	private String member_join;
-	private int member_level;
+	private double member_level;
 	private int member_jointype;
 	private String member_role;
 }
